Move all list methods to different controllers.
